fix(planungshilfen): correct visualization of gaps between months
The previous implementation aggregated all months with the same option
into a single total. This caused a bug where non-consecutive selections
were incorrectly shown as continuous.

For example, if a parent chose Basis in the 1st and 13th month, the
visualization displayed it as if Basis were chosen in the 1st and
2ndmonth instead.