refactor: expose setzePlanZurueck as imperative handle
We are adding more and more functionality to the header of the planer,
which is not really related to the planer component.

I started exposing the functionality of the planer in this commit with
useImperativeHandle, with the vision in mind that the planer will soon
no longer contain the anleitung or the header buttons.