feat: synchronize referrer metrics between matomo and metabase
We launched a campaign with flyers referring to the elterngeldrechner
and added qr codes to two pages of the paper application. All qr codes
point to the elterngeldrechner and include a `source` query parameter.

This commit synchronizes referrer data from matomo to metabase to
provide insights into the impact of these changes.