fix: line wrapping of einkommen on smaller screens
The icon for the Einkommen did wrap on smaller screens. This change sets
the min-width of the grid column and also makes the option text scale.
The change is not a guaranteed fix for all cases; for example, an
Einkommen of 500.000 Euro would still cause line wrapping, but it is
good enough for most cases, like Einkommen up to 99.999 Euro.