feat(tracking): add geplante monate and vatermonate to data layer
One of the previous refactorings accidentally removed the variables
which are still needed, at list till the end of the current campaign.
During the re-add I refactored the filterting of months so that all
three functions use the same counting functions. The common filtering
method solves a bug where in the prev version the 'Kein Elterngeld' was
counted as planned month in one function but not in another.