fix: navigate to datenuebernahme with only mutterschutz selecte
In case the user didn't change the plan, the plan on the PlanerPage
remains undefined, which prevents the user from navigating to the
DatenuebernahmeAntragPage.
The root issue is that the plan is handled by the Planer
as well as the PlanerPage.
To resolve the before mentioned issue, as an initial plan with only
two months of Mutterschutz is also considered a valid plan,
the Planer component now passes the plan to the PlanerPage
when clicking on planInAntragUebernehmen.
Another option, which we didn't continue further, would have been
to pass the PlanerPage shell to the Planer as slots and handle
the plan from there.