feat: add package mutterschutzrechner
This commit adds the package mutterschutzrechner,
which calculates the start and end date of the
mutterschutz by entering only the expected date
of birth of both the expected date and the
actual date and, if applicable, also a reason
for a longer time period after the birth.
It includes all the necessary calculations, types
and tests for this package.