feat: add package bemessungszeitraumrechner with standard calculations
This commit adds a first part of the package
bemessungszeitraumrechner. It includes all the
necessary types and the calculations for the
preliminary time periods for a first grasp of
the relevant time periods for the Bemessungszeitraum
as well as the calculations for the standard cases,
where parents have no time periods of exclusion which
would change the time period of the Bemessungszeitraum.

It is still WIP, meaning that at least tests for
those calculations and the exact calculations
are to be added.