feat: change new dates to utc dates and add tests
This commit changes all new Dates to
utc dates to prevent problems with
time zones etc.
Furthermore it adds test cases for the
exact calculations and the calculation
of the relevant time periods.
Tests for the calculations of the
Bemessungszeitraum without regard of
the auszuklammernde zeitraume are still
to be done.