feat: changes to mutterschutzrechner logic and tests
This commit changes the logic in which the
startdatum is calculated by setting the geburtsdatum
as the startdatum if the birth was more than six
weeks prior to the expected date.
It also changes the calculation of the
additional period after the birth by adding all
days for which the actual birth took place before the
expected birth.
Furthermore it refactors the tests to use test.each with
defined test cases.
And finally there are some small changes:
- add relevant paragraphs of the law to the docs
- add branded type for Mutterschutz
- make constants self-documenting
- add one constant and rename one