feat: simplify verlaengerungsGrund for Mutterschutzrechner
This commit simplifies the logic for a
Verlängerungsgrund as it doesn't expect a
certain reason anymore but simply takes a
boolean as argument to calculate the longer
time period after the birth.
Additionally this commit moves the
constants to the berechneMutterschutz file
and uses a function to create the milliseconds
of the weeks for the calculations.
Lastly it removes the use of the branded type.