chore: disable automatic script execution in npm
The shai-hulud-2 incident a couple of days ago resulted
in numerous repositories being compromised by exploiting
nodes lifecycle hooks.

For now, the safest measure on our side is to completely
disable npm’s automatic script execution. The downside is
that we must now prepare the original-rechner explicitly, since
the npm cli currently does not provide a mechanism to
selectively allow trusted packages or scripts when
script execution is globally disabled.

https://github.com/npm/cli/issues/8781