test: fix failing playwright test due to cookie banner
The familienportal, and shell of the elterngeldrechner,
has introduced a change that makes it mandatory to
interact with the cookie banner before using the page.

The focus of a users device will automatically move
to the cookie banner if they try to interact with the
elterngeldrechner without accepting or declining
the cookie banner.